feat(tenant): 实现租户消息同步机制
- 新增租户消息监听器和发布者- 实现租户数据同步逻辑
- 优化 Redis 配置和序列化方式
- 新增租户相关的 mapper 和 SQL

+++ spring-support-mybatis-starter/src/main/java/com/chua/starter/mybatis/pojo/SysTenantBase.java
@@ -21,5 +21,5 @@ public class SysTenantBase extends SysBase{
     @Column(defaultValue = "0", comment = "租户ID", refresh = true)
     @Schema(description = "租户ID")
     @ApiModelProperty("租户ID")
-    private Long sysTenantId;
+    private Integer sysTenantId;
 }
